requirements.txt
streamlit
pymongo
sentence-transformers
bcrypt
python-dotenv
PyJWT
